+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added:
+
+- Support for more OneStamp shapes
+
 ## [1.1.0] - 2026-05-25
 
 ### Added:
@@ -27,4 +33,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [1.1.0]: https://github.com/Radiokot/press-cut-app/compare/1.0.1(6)...1.1.0(7)
 
-[Unreleased]: https://github.com/Radiokot/press-cut-app/compare/1.0.1(6)...HEAD
+[Unreleased]: https://github.com/Radiokot/press-cut-app/compare/1.1.0(7)...HEAD